requirements.txt
streamlit
google-generativeai
python-dotenv
python-docx
pdfplumber
langchain